Added the distinction between a prediction and a rating.
A prediction would be an initial score generated by the system, and a rating would be something to overwrite that. This includes the ability to revert to the prediction when a rating is cleared, and allowing the prediction and rating colors to be different. Also added the ability to rate only whole values, rather than any value.